split FGattInfoFeatureApi
