+++ components/bridge/connection/feature/common/api/src/commonMain/kotlin/com/flipperdevices/bridge/connection/feature/common/api/FDeviceFeature.kt
@@ -13,7 +13,8 @@ enum class FDeviceFeature {
     STORAGE_INFO,
     GET_INFO,
     ALARM,
-    DEVICE_COLOR
+    DEVICE_COLOR,
+    GATT_INFO,
 }
 
 @Retention(AnnotationRetention.RUNTIME)
