+++ components/bridge/connection/feature/device-color/impl/src/commonMain/kotlin/com/flipperdevices/bridge/connection/feature/devicecolor/impl/api/FDeviceColorFeatureFactoryImpl.kt
@@ -16,7 +16,7 @@ import com.squareup.anvil.annotations.ContributesMultibinding
 import kotlinx.coroutines.CoroutineScope
 import javax.inject.Inject
 
-
+// todo move to shared const
 private val API_SUPPORTED_GET_REQUEST = SemVer(
     majorVersion = 0,
     minorVersion = 14
