+++ components/bridge/connection/feature/getinfo/api/src/commonMain/kotlin/com/flipperdevices/bridge/connection/feature/getinfo/api/FGattInfoFeatureApi.kt
@@ -1,0 +1,9 @@
+package com.flipperdevices.bridge.connection.feature.getinfo.api
+
+import com.flipperdevices.bridge.connection.feature.common.api.FDeviceFeatureApi
+import com.flipperdevices.bridge.connection.feature.getinfo.model.FGattInformation
+import kotlinx.coroutines.flow.StateFlow
+
+interface FGattInfoFeatureApi : FDeviceFeatureApi {
+    fun getGattInfoFlow(): StateFlow<FGattInformation>
+}
