+++ components/bridge/connection/feature/getinfo/api/src/commonMain/kotlin/com/flipperdevices/bridge/connection/feature/getinfo/api/FGetInfoFeatureApi.kt
@@ -1,16 +1,12 @@
 package com.flipperdevices.bridge.connection.feature.getinfo.api
 
 import com.flipperdevices.bridge.connection.feature.common.api.FDeviceFeatureApi
-import com.flipperdevices.bridge.connection.feature.getinfo.model.FGattInformation
 import com.flipperdevices.bridge.connection.feature.getinfo.model.FGetInfoApiGroup
 import com.flipperdevices.bridge.connection.feature.getinfo.model.FGetInfoApiProperty
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
 
 interface FGetInfoFeatureApi : FDeviceFeatureApi {
     suspend fun get(property: FGetInfoApiProperty): Result<String>
 
     fun get(group: FGetInfoApiGroup): Flow<Pair<FGetInfoApiProperty, String>>
-
-    fun getGattInfoFlow(): MutableStateFlow<FGattInformation>
 }
